[GraphMaker] Fix scan issues
